fp7 files
Check the notes that i took when i talked to the teacher. Ask me if you don't understand. First we need to model the database and create mockups for the pages and only then code it.

I also uploaded some stuff from the exercise. Basically it is copied from the slides, except the productController.js which in slides is not compatible with new mongoose.